# docs/openapi/swagger-spec.yaml
openapi: 3.0.0
paths:
  /portfolio/{id}/header:
    get:
      operationId: PortfolioController_getHeader
      summary: Returns the header of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeaderDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/about:
    get:
      operationId: PortfolioController_getAbout
      summary: Returns the about of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AboutDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/education:
    get:
      operationId: PortfolioController_getEducation
      summary: Returns the education of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EducationPortfolioDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/work-experience:
    get:
      operationId: PortfolioController_getWorkExperience
      summary: Returns the work experience of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkExperiencePortfolioDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/skills:
    get:
      operationId: PortfolioController_getSkills
      summary: Returns the skills of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillPortfolioDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/projects:
    get:
      operationId: PortfolioController_getProjects
      summary: Returns the projects of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectPortfolioDto'
      tags:
        - portfolio
      security:
        - bearer: []
  /portfolio/{id}/footer:
    get:
      operationId: PortfolioController_getFooter
      summary: Returns the footer of a user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FooterPortfolioDto'
      tags:
        - portfolio
      security:
        - bearer: []
...
info:
  title: Portfolio API
  description: The OpenAPI of the Portoflio API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    HeaderDto:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        description:
          type: string
        navBar:
          type: array
          items:
            type: string
        iconList:
          type: array
          items:
            type: string
      required:
        - name
        - role
        - description
        - navBar
        - iconList
    AboutDto:
      type: object
      properties:
        paragraphs:
          type: array
          items:
            type: string
      required:
        - paragraphs
    EducationPortfolioDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        university:
          type: string
        date:
          type: string
        universityLink:
          type: string
      required:
        - id
        - name
        - university
        - date
        - universityLink
    WorkExperiencePortfolioDto:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        company:
          type: string
        companyLink:
          type: string
        date:
          type: string
        description:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - role
        - company
        - companyLink
        - date
        - description
        - links
        - tags
    SkillPortfolioDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        progress:
          type: number
      required:
        - id
        - name
        - progress
    ProjectPortfolioDto:
      type: object
      properties:
        id:
          type: string
        date:
          type: number
        title:
          type: string
        subtitle:
          type: string
        projectLink:
          type: string
        imageLink:
          type: string
        description:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - date
        - title
        - subtitle
        - projectLink
        - imageLink
        - description
        - links
        - tags
    FooterPortfolioDto:
      type: object
      properties:
        paragraphs:
          type: array
          items:
            type: string
      required:
        - paragraphs
...